+++ NEWS.md
@@ -3,7 +3,8 @@ RRTMGP.jl Release Notes
 
 main
 ------
-- Remove pade approximation. PR [#563](https://github.com/CliMA/RRTMGP.jl/pull/563)
+Update cloud optics to the latest version of rrtmgp-data.
+PR [#562](https://github.com/CliMA/RRTMGP.jl/pull/562)
 
 v0.19.1
 -----